change init_idx to INT32 in transformer_test
test=develop

+++ paddle/fluid/inference/tests/api/analyzer_transformer_tester.cc
@@ -147,7 +147,7 @@ void PrepareInputs(std::vector<PaddleTensor> *input_slots, DataRecord *data,
 
   init_idx.name = "init_idx";
   init_idx.shape.assign({batch_size});
-  init_idx.dtype = PaddleDType::INT64;
+  init_idx.dtype = PaddleDType::INT32;
   TensorAssignData<int64_t>(&init_idx, one_batch.init_idx);
 
   trg_src_attn_bias.name = "trg_src_attn_bias";
